refactor(ai): holistic cleanup — shared json defaults, per-step spans, structured termination logs
- Extract AiJsonDefaults.Serializer shared across AgentToolDispatcher and ChatAgentRunSink
- Add per-step ai.agent.step activity spans with kind/token/tool_count tags in both non-streaming and streaming loops
- Log structured Info on LoopBreak and MaxStepsExceeded termination sites
- Add tool_name tag to LoopBreaks counter (parsed from TerminationReason)
- Extract StepBudgetExceededMessage const in ChatExecutionService (was duplicated 4×)
- Drop stray async on OnAssistantMessageAsync (no await in body)
- Add XML docs on ThinkOnly enum value and AgentRunSinkEvents StepIndex note
- Widen AgentRunResult token totals to long; cast to int at FinalizeTurnAsync call sites
- Upgrade Ollama tool-strip log from Info to Warning
- Add defensive comment on Cancelled branch in ExecuteStreamAsync

+++ boilerplateBE/src/modules/Starter.Module.AI/Application/Services/Runtime/AgentRunResult.cs
@@ -4,8 +4,8 @@ internal sealed record AgentRunResult(
     AgentRunStatus Status,
     string? FinalContent,
     IReadOnlyList<AgentStepEvent> Steps,
-    int TotalInputTokens,
-    int TotalOutputTokens,
+    long TotalInputTokens,
+    long TotalOutputTokens,
     string? TerminationReason);
 
 internal enum AgentRunStatus
